+++ src/main/scala/com/github/wladox/model/XwaySegDir.scala
@@ -3,6 +3,6 @@ package com.github.wladox.model
 /**
   * Created by root on 21.01.17.
   */
-case class XwaySegDir(xWay: Int, segment:Int, direction:Int) {
+case class XwaySegDir(xWay:Byte, segment:Byte, direction:Byte) {
   override def toString: String = s"$xWay.$segment.$direction"
 }
